server/requirements.txt
flask==3.0.0
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.6.0
flask_bcrypt==1.0.1
pymysql==1.1.0
Pillow==10.1.0
